requirements.txt
# Core
Django>=5.0,<6.0
djangorestframework>=3.15.0
python-dotenv>=1.0.0
httpx>=0.27.0
phonenumbers>=8.13.0

# Retell signature verification (official SDK)
retell-sdk>=5.0.0

# Production DB (optional; SQLite used when DATABASE_URL unset)
psycopg[binary]>=3.2.0
